fix: added circular import validation
