+++ src/auth/signature-verifier.ts
@@ -11,10 +11,9 @@ import {RemoteJWKSetOptions} from 'jose/dist/types/jwks/remote';
 import {debug} from '../debug/index.js';
 import {AuthError, AuthErrorCode} from './error.js';
 import {useEmulator} from './firebase.js';
-import {VerifyOptions, getPublicCryptoKey, verify} from './jwt/verify.js';
+import {getPublicCryptoKey, verify} from './jwt/verify.js';
+import {VerifyOptions} from './types.js';
 import {isNonNullObject, isURL} from './validator.js';
-
-export const ALGORITHM_RS256 = 'RS256' as const;
 
 export type PublicKeys = {[key: string]: string};
