+++ src/next/cookies/parser/ObjectCookiesProvider.ts
@@ -1,4 +1,4 @@
-import {CookiesObject} from '../index.js';
+import {CookiesObject} from '../types.js';
 
 export class ObjectCookiesProvider {
   constructor(private cookies: CookiesObject) {}
