+++ src/next/cookies/verification.ts
@@ -1,0 +1,56 @@
+import type {ReadonlyRequestCookies} from 'next/dist/server/web/spec-extension/adapters/request-cookies';
+import type {RequestCookies} from 'next/dist/server/web/spec-extension/cookies';
+import type {NextResponse} from 'next/server';
+import {CookiesObject} from './types.js';
+
+const INTERNAL_VERIFIED_TOKEN_COOKIE_NAME =
+  'x-next-firebase-auth-edge-verified';
+const INTERNAL_VERIFIED_TOKEN_COOKIE_VALUE = 'true';
+
+export function areCookiesVerifiedByMiddleware(
+  cookies: RequestCookies | ReadonlyRequestCookies
+) {
+  return (
+    cookies.get(INTERNAL_VERIFIED_TOKEN_COOKIE_NAME)?.value ===
+    INTERNAL_VERIFIED_TOKEN_COOKIE_VALUE
+  );
+}
+
+export function isCookiesObjectVerifiedByMiddleware(cookies: CookiesObject) {
+  return (
+    cookies[INTERNAL_VERIFIED_TOKEN_COOKIE_NAME] ===
+    INTERNAL_VERIFIED_TOKEN_COOKIE_VALUE
+  );
+}
+
+export function removeInternalVerifiedCookieIfExists(
+  cookies: RequestCookies | ReadonlyRequestCookies
+) {
+  if (cookies.get(INTERNAL_VERIFIED_TOKEN_COOKIE_NAME)?.value) {
+    cookies.delete(INTERNAL_VERIFIED_TOKEN_COOKIE_NAME);
+  }
+}
+
+export function markCookiesAsVerified(
+  cookies: RequestCookies | ReadonlyRequestCookies
+) {
+  cookies.set(
+    INTERNAL_VERIFIED_TOKEN_COOKIE_NAME,
+    INTERNAL_VERIFIED_TOKEN_COOKIE_VALUE
+  );
+}
+
+export function wasResponseDecoratedWithModifiedRequestHeaders(
+  response: NextResponse
+) {
+  const cookie = response.headers.get('cookie');
+  const middlewareRequestCookie = response.headers.get(
+    'x-middleware-request-cookie'
+  );
+
+  return (
+    cookie?.includes(INTERNAL_VERIFIED_TOKEN_COOKIE_NAME) ||
+    middlewareRequestCookie?.includes(INTERNAL_VERIFIED_TOKEN_COOKIE_NAME) ||
+    false
+  );
+}
